convert engine to package
