+++ bzt/engine/__init__.py
@@ -1,0 +1,3 @@
+from .names import EXEC, TAURUS_ARTIFACTS_DIR, SETTINGS
+from .engine import Engine, Configuration, FileLister, Scenario, HavingInstallableTools, Singletone, SelfDiagnosable
+from .modules import Provisioning, Reporter, Service, Aggregator, EngineModule, ScenarioExecutor
